try and run SSR on vercel
